+++ app/src/main/java/woowacourse/shopping/data/cart/CartDao.kt
@@ -4,6 +4,7 @@ import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
+import woowacourse.shopping.data.product.ProductEntity
 
 @Dao
 interface CartDao {
