+++ app/src/main/java/woowacourse/shopping/data/cart/CartDatabase.kt
@@ -1,9 +1,10 @@
-package woowacourse.shopping.data.product
+package woowacourse.shopping.data.cart
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import woowacourse.shopping.data.product.ProductEntity
 
 @Database(entities = [ProductEntity::class], version = 1)
 abstract class CartDatabase : RoomDatabase() {
